Make type conversion a part of arithmetics library
