+++ lib/src/atom/serde.rs
@@ -21,36 +21,3 @@ pub enum Error {
 }
 
 pub type Result = std::result::Result<(), Error>;
-
-#[derive(Default)]
-pub(crate) struct BoolSerializer {
-    value: Option<bool>,
-}
-
-impl Serializer for BoolSerializer {
-    fn serialize_bool(&mut self, v: bool) -> Result {
-        self.value = Some(v);
-        Ok(())
-    }
-}
-
-use std::convert::TryInto;
-
-pub struct AsPrimitive<'a> {
-    atom: &'a super::Atom
-}
-
-impl<'a> AsPrimitive<'a> {
-    pub fn from_atom(atom: &'a super::Atom) -> Self {
-        Self{ atom }
-    }
-
-    pub fn as_bool(self) -> Option<bool> {
-        TryInto::<&dyn super::GroundedAtom>::try_into(self.atom).ok()
-            .map(|gnd| {
-                let mut serializer = BoolSerializer::default();
-                let _ = gnd.serialize(&mut serializer);
-                serializer.value
-            }).flatten()
-    }
-}
